Move Mamba2 into the SSM layer and fix Nemotron dt_bias init
Address review feedback on the Nemotron variant:

- Fix NemotronMamba2Mixer dt_bias initialization. The mixer previously
  discarded time_step_max/time_step_floor and initialized dt_bias to
  ones, so softplus(dt_bias) started at ~1.31 (>13x the configured
  timestep max). Now match the Nemotron-H/Mamba2 init: sample dt
  log-uniformly in [time_step_min, time_step_max], clamp by
  time_step_floor, and store the inverse-softplus. Add a regression test
  asserting initialized effective timesteps stay in range.

- Move NemotronMamba2Config/Mixer (plus NemotronRMSNormGated and the
  chunk-scan helpers) into the reusable SSM sequence-mixer layer
  nn/attention/recurrent.py, registered as a SequenceMixer
  ("nemotron_mamba2") mirroring GatedDeltaNet, and re-exported from the
  attention package. nemotron.py now only assembles the variant. Mamba2
  unit tests move to recurrent_test.py.
